Combine DiffUtilCallbacks
