+++ app/src/main/java/com/cornellappdev/volume/util/DiffUtilCallbacks.kt
@@ -1,15 +1,46 @@
-package com.cornellappdev.volume
+package com.cornellappdev.volume.util
 
 import androidx.annotation.Nullable
 import androidx.recyclerview.widget.DiffUtil
+import com.cornellappdev.volume.models.Article
 import com.cornellappdev.volume.models.Publication
+
+/**
+ * DiffUtil makes the replacement of articles in RecyclerViews faster
+ * by comparing the strict difference between oldArticles and newArticles
+ * and calling notifyDataSetChanged for the items accordingly.
+ */
+class DiffUtilCallbackArticle(private val oldArticles: List<Article>, private val newArticles: List<Article>) :
+    DiffUtil.Callback() {
+
+    override fun getOldListSize(): Int {
+        return oldArticles.size
+    }
+
+    override fun getNewListSize(): Int {
+        return newArticles.size
+    }
+
+    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+        return oldArticles[oldItemPosition].id === newArticles[newItemPosition].id
+    }
+
+    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+        return oldArticles[oldItemPosition] == newArticles[newItemPosition]
+    }
+
+    @Nullable
+    override fun getChangePayload(oldItemPosition: Int, newItemPosition: Int): Any? {
+        return super.getChangePayload(oldItemPosition, newItemPosition)
+    }
+}
 
 /**
  * DiffUtil makes the replacement of publications in RecyclerViews faster
  * by comparing the strict difference between oldPublications and newPublications
  * and calling notifyDataSetChanged for the items accordingly.
  */
-class DiffUtilCallback(
+class DiffUtilCallbackPublication(
     private val oldPublications: List<Publication>,
     private val newPublications: List<Publication>
 ) :
